Moved frequency-model and lemmatizer to cltk.lemmatize.latin.unsupervised.py
